This commit was manufactured by cvs2git to create tag 'auto-ac-alpine- 1_00_6-1'.
Sprout from master 2007-12-21 10:09:25 UTC Paweł Gołaszewski <blues@pld-linux.org> '- fixed for 1.00, bit extended'
Cherrypick from master 2007-12-26 14:13:16 UTC Jakub Bogusz <qboosh@pld-linux.org> '- removed extra space':
    alpine.desktop -> 1.3
Cherrypick from unlabeled-1.28.2 2008-01-02 10:20:14 UTC Paweł Gołaszewski <blues@pld-linux.org> '- md':
    alpine.spec -> 1.28.2.13
Delete:
    alpine-fixhome.patch

+++ alpine.spec
@@ -2,11 +2,11 @@
 # - backport man-pages from pine.spec
 # - separate package with tcl web-frontend
 # - fix as-needed
-Summary:	University of Washington Alpine mail user agent
-Summary(pl.UTF-8):	Klient pocztowy Alpine z Uniwersytetu w Waszyngtonie
+Summary:	University of Washington Pine mail user agent
+Summary(pl.UTF-8):	Klient pocztowy Pine z Uniwersytetu w Waszyngtonie
 Name:		alpine
 %define		ver		1.00
-%define		patchlevel	3
+%define		patchlevel	6
 Version:	%{ver}.%{patchlevel}
 Release:	1
 Epoch:		1
@@ -16,7 +16,7 @@ Group:		Applications/Mail
 #Source0:	ftp://ftp.cac.washington.edu/alpine/%{name}-%{version}.tar.gz
 # Source with applied patches from http://staff.washington.edu/chappa/alpine/
 Source0:	http://staff.washington.edu/chappa/alpine/patches/alpine-%{ver}/%{name}-%{ver}_%{patchlevel}.tar.gz
-# Source0-md5:	3937652526da1afe6310e05650ace80c
+# Source0-md5:	2de2a8264512e557898769a286494819
 Source1:	pico.desktop
 Source2:	%{name}.desktop
 Source3:	%{name}.png
@@ -27,23 +27,21 @@ Patch3:		%{name}-quote.patch
 Patch4:		%{name}-fhs.patch
 Patch5:		%{name}-segfix.patch
 Patch6:		%{name}-libc-client.patch
-Patch7:		%{name}-fixhome.patch
-Patch8:		%{name}-ssl.patch
-Patch9:		%{name}-no_1777_warning.patch
-Patch10:	%{name}-home_etc.patch
+Patch7:		%{name}-ssl.patch
+Patch8:		%{name}-no_1777_warning.patch
+Patch9:		%{name}-home_etc.patch
 URL:		http://www.washington.edu/alpine/
 BuildRequires:	autoconf
 BuildRequires:	automake
 BuildRequires:	home-etc-devel
-BuildRequires:	krb5-devel
 BuildRequires:	ncurses-devel
-BuildRequires:	openldap-devel >= 2.4.6
+BuildRequires:	openldap-devel
 BuildRequires:	openssl-devel
 BuildRequires:	pam-devel
 # Only for web-frontend:
 #BuildRequires:	tcl-devel
 Obsoletes:	pine
-Provides:	pine = 4.99
+Provides:	pine = 5.00
 Suggests:	aspell
 BuildRoot:	%{tmpdir}/%{name}-%{version}-root-%(id -u -n)
 
@@ -75,7 +73,7 @@ Summary:	Simple text editor in the style of the Pine Composer
 Summary(pl.UTF-8):	Prosty edytor tekstowy w stylu alpine
 Summary(pt_BR.UTF-8):	Editor de textos para terminal simples e fácil de usar
 Group:		Applications/Editors
-Provides:	pico = 4.99
+Provides:	pico = 5.00
 
 %description -n pico
 Pico is a simple, display-oriented text editor based on the Alpine
@@ -99,7 +97,7 @@ Summary:	Simple file system browser in the style of the Alpine Composer
 Summary(pl.UTF-8):	Prosta przeglądarka plików w stylu composera alpine
 Summary(pt_BR.UTF-8):	Navegador de sistemas de arquivos no estilo do compositor do Alpine
 Group:		Applications/Shells
-Provides:	pilot = 4.99
+Provides:	pilot = 5.00
 
 %description -n pilot
 Pilot is a simple, display-oriented file system browser based on the
@@ -108,7 +106,7 @@ at the bottom of the screen, and context-sensitive help is provided.
 
 %description -n pilot -l pl.UTF-8
 Pilot jest prostą, zorientowaną na wyświetlanie przeglądarką plików w
-stylu composera pine. Podobnie jak w alpine polecenia są wyświetlane
+stylu composera pine. Podobnie jak w alpine polecenia sa wyświetlane
 na dole ekranu oraz jest dostępna pomoc kontekstowa.
 
 %description -n pilot -l pt_BR.UTF-8
@@ -128,7 +126,6 @@ ajuda de acordo com o contexto está disponível.
 %patch7 -p1
 %patch8 -p1
 %patch9 -p1
-%patch10 -p1
 
 %build
 rm -f libtool missing
@@ -144,7 +141,6 @@ rm -f libtool missing
 	--with-simple-spellcheck=aspell \
 	--with-system-pinerc=%{alpineconfdir}/%{name}.conf \
 	--with-system-fixed-pinerc=%{alpineconfdir}/%{name}.conf.fixed \
-	--with-krb5-dir=%{_prefix} \
 	--with-ldap-dir=%{_prefix} \
 	--with-system-mail-directory=/var/mail \
 	--with-c-client-target=slx \
@@ -181,6 +177,12 @@ cat <<EOF > $RPM_BUILD_ROOT%{alpineconfdir}/alpine.conf.fixed
 # comments at the top of %{alpineconfdir}/alpine.conf
 
 EOF
+
+%post
+if [ -f "%{alpineconfdir}/alpine.conf" -a -f "%{alpineconfdir}/alpine.conf.rpmnew" ]; then
+	mv %{alpineconfdir}/alpine.conf %{alpineconfdir}/alpine.conf.backup
+	alpine -P %{alpineconfdir}/alpine.conf.backup -conf > /etc/alpine/alpine.conf || exit 0
+fi
 
 %clean
 rm -rf $RPM_BUILD_ROOT
